feat(table/cells): refactor
