+++ src/utils/index.js
@@ -2,3 +2,4 @@ export * from './function'
 export * from './string'
 export * from './tabs'
 export * from './time'
+export * from './color'
